refactor: share Spark row materialization across harnesses
This change extracts writeRows(), coerce(), cleanup() into a public
trait so that different workload harnesses can reuse these functions
instead of reimplementing them. This change also fixes some compilation
issues which allow builds to run on Scala 2.12
